ChaiplusV2

+++ Algo/pgcd.c
@@ -1,19 +1,19 @@
 #include <stdio.h>
 
-int main(){
-    int nb1 = 36, nb2 = 90;
+long long main(){
+    long long nb1 = 36, nb2 = 90;
     printf("Valeur nb1 et nb2:");
-    scanf("%d %d", &nb1, &nb2);
+    scanf("%lld %lld", &nb1, &nb2);
 
-    int copy_nb1 = nb1, copy_nb2 = nb2;
+    long long copy_nb1 = nb1, copy_nb2 = nb2;
 
-    int t;
+    long long t;
     while (copy_nb2 != 0){
        t = copy_nb2; 
        copy_nb2 = copy_nb1 % copy_nb2; 
        copy_nb1 = t;
     }
     
-    printf("PGDC(%d, %d) = %d\n", nb1, nb2, copy_nb1);
+    printf("PGDC(%lld, %lld) = %lld\n", nb1, nb2, copy_nb1);
     return 0;
 }
